requirements.txt
streamlit
firebase-admin
groq
python-dotenv
PyMuPDF

python-pptx
paddleocr
paddlepaddle

torch
transformers
pillow
accelerate